TRESTLE-757: Refactor auth reducers and state into their own file.
