+++ trestle-server/src/main/webapp/src/app/navigation/top-nav/top-nav.component.ts
@@ -2,7 +2,8 @@ import {Component, EventEmitter, OnInit, Output} from '@angular/core';
 import {Observable} from 'rxjs';
 import {TrestleUser} from '../../user/trestle-user';
 import {select, Store} from '@ngrx/store';
-import {selectUserFromUser, State} from '../../reducers';
+import {State} from '../../reducers';
+import {selectUserFromUser} from "../../reducers/auth.reducers";
 
 @Component({
   selector: 'top-nav',
