+++ trestle-server/src/main/webapp/src/app/reducers/index.ts
@@ -1,71 +1,13 @@
-import {ActionReducerMap, createReducer, createSelector, MetaReducer, on} from '@ngrx/store';
-import {TrestleUser} from '../user/trestle-user';
-import {login, loginFailure, loginSuccess, logout} from '../actions/auth.actions';
+import {ActionReducerMap, MetaReducer} from '@ngrx/store';
 import {storageMetaReducer} from './storage.metareducer';
-
-export interface UserState {
-  user: TrestleUser;
-  userError: Error;
-  userToken: string;
-}
-
-const initialUserState: UserState = {
-  user: null,
-  userError: null,
-  userToken: ''
-};
+import {authReducer, UserState} from "./auth.reducers";
 
 export interface State {
   user: UserState;
-}
-
-const _authReducer = createReducer(initialUserState, on(login, state => {
-    console.log('Trying to login');
-    return state;
-  }),
-  on(loginSuccess, (state, {user, token}) => ({
-    userError: null,
-    user,
-    userToken: token
-  })),
-  on(loginFailure, (state, {error}) => ({
-    user: null,
-    userError: error,
-    userToken: ''
-  })),
-  on(logout, (state) => ({
-    ...state,
-    user: null,
-    userToken: ''
-  })));
-
-
-export function authReducer(state: UserState = initialUserState, action): UserState {
-  return _authReducer(state, action);
 }
 
 export const reducers: ActionReducerMap<State> = {
   user: authReducer
 };
 
-
 export const metaReducers: MetaReducer<State>[] = [storageMetaReducer];
-
-
-export const selectUserState = (state: State) => state.user;
-export const selectUser = (state: UserState) => state.user;
-
-
-export const selectUserFromUser = createSelector(
-  selectUserState,
-  selectUser);
-
-export const selectErrorFromUser = createSelector(
-  selectUserState,
-  (state) => state.userError
-);
-
-export const selectTokenFromUser = createSelector(
-  selectUserState,
-  (state) => state.userToken
-);
